flredecomp fix / load balancing fix.
When flredecomping an extruded mesh, Fluidity used to weight the InputMesh
nodes according by # of nodes in a column, with weights normalised between 0-1.

If extruded meshes had variable numbers of layers, this could end up with a 
load imbalance of 5:1 or higher. 

zoltan_cb_get_owned_nodes() now simply uses the # column nodes as a weight,
which gives well balanced partitions.