The now-functional compile-time optimisations (CTO) #defines generator.
How it works:
	./configure --enable-cto=path/to/simulation/sim_config.flml

Where sim_config.flml is the FLML file used for simulation configuration.

This generates an additional include file, include/compile_opt_defs.h, which
includes #defines for ndim, nloc, ngi, etc. relevant to the FLML file.

It also adds '#define USE_CTO 1' to include/confdefs.h, so including it
can be enabled.

NOTE:
----
Eventually this will produce a local fluidity executable in the simulation
directory, but that's for later.